:recycle: api: Move getSessionCertificationReports injection to course
