+++ api/src/certification/course/application/certification-reports-controller.js
@@ -1,4 +1,4 @@
-import { usecases } from '../../shared/domain/usecases/index.js';
+import { usecases } from '../../course/domain/usecases/index.js';
 import * as certificationReportSerializer from '../../shared/infrastructure/serializers/jsonapi/certification-report-serializer.js';
 
 const getCertificationReports = async function (request, h, dependencies = { certificationReportSerializer }) {
